Adjust the scan resolution link
