+++ app/src/main/java/connect/ui/activity/contact/ScanAddFriendActivity.java
@@ -29,10 +29,11 @@ import connect.db.SharedPreferenceUtil;
 import connect.ui.activity.R;
 import connect.ui.activity.home.bean.MsgNoticeBean;
 import connect.ui.activity.login.bean.UserBean;
-import connect.ui.activity.wallet.support.ScanUrlAnalysisUtil;
 import connect.ui.base.BaseScanActivity;
 import connect.utils.ActivityUtil;
 import connect.utils.ProgressUtil;
+import connect.utils.scan.ResolveScanUtil;
+import connect.utils.scan.ResolveUrlUtil;
 import connect.utils.system.SystemUtil;
 import connect.view.ScanBgView;
 import connect.view.album.entity.ImageInfo;
@@ -78,7 +79,7 @@ public class ScanAddFriendActivity extends BaseScanActivity {
 
     private ScanAddFriendActivity mActivity;
     private UserBean userBean;
-    private ScanUrlAnalysisUtil analysisUtil;
+    private ResolveScanUtil resolveScanUtil;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -100,17 +101,17 @@ public class ScanAddFriendActivity extends BaseScanActivity {
         scanImg.setImageBitmap(bitmap);
         bottomScanImg.setImageBitmap(bitmap);
         addressTv.setText(userBean.getAddress());
-        analysisUtil = new ScanUrlAnalysisUtil(mActivity);
+        resolveScanUtil = new ResolveScanUtil(mActivity);
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
     public void onEventMainThread(MsgNoticeBean notice) {
-        analysisUtil.showMsgTip(notice, "scan");
+        new ResolveUrlUtil(mActivity).showMsgTip(notice, true);
     }
 
     @Override
     public void scanCall(String value) {
-        analysisUtil.analysisUrl(value);
+        resolveScanUtil.analysisUrl(value);
     }
 
     @OnClick(R.id.right_img)
@@ -230,7 +231,7 @@ public class ScanAddFriendActivity extends BaseScanActivity {
             ProgressUtil.getInstance().dismissProgress();
             switch (msg.what){
                 case PARSE_BARCODE_SUC:
-                    analysisUtil.analysisUrl((String) msg.obj);
+                    resolveScanUtil.analysisUrl((String) msg.obj);
                     break;
             }
