--- app/src/main/java/connect/ui/activity/login/bean/WebOpenBean.java
@@ -1,45 +1,0 @@
-package connect.ui.activity.login.bean;
-
-/**
- * Created by Administrator on 2017/2/9.
- */
-
-public class WebOpenBean {
-
-    String address;
-    String token;
-    String amount;
-    String type;
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-
-    public String getAmount() {
-        return amount;
-    }
-
-    public void setAmount(String amount) {
-        this.amount = amount;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-}
