+++ app/src/main/java/connect/ui/activity/login/presenter/StartPresenter.java
@@ -25,12 +25,9 @@ import java.util.Random;
 import connect.db.SharedPreferenceUtil;
 import connect.ui.activity.login.bean.StartImagesBean;
 import connect.ui.activity.login.bean.UserBean;
-import connect.ui.activity.login.bean.WebOpenBean;
 import connect.ui.activity.login.contract.StartContract;
-import connect.ui.activity.wallet.support.ScanUrlAnalysisUtil;
 import connect.ui.base.BaseApplication;
 import connect.utils.system.SystemDataUtil;
-import connect.utils.system.SystemUtil;
 import connect.utils.UriUtil;
 import connect.utils.glide.GlideUtil;
 import connect.utils.glide.OnDownloadTarget;
@@ -196,13 +193,16 @@ public class StartPresenter implements StartContract.Presenter{
         }
     }
 
+    /**
+     * Save url with open the App
+     * @param activity
+     */
     private void openFromWeb(Activity activity){
         Intent i_getvalue = activity.getIntent();
         String action = i_getvalue.getAction();
         if(Intent.ACTION_VIEW.equals(action)){
             Uri uri = i_getvalue.getData();
-            WebOpenBean webOpenBean = ScanUrlAnalysisUtil.WebOpenAppData(uri);
-            SharedPreferenceUtil.getInstance().putValue(SharedPreferenceUtil.WEB_OPEN_APP,new Gson().toJson(webOpenBean));
+            SharedPreferenceUtil.getInstance().putValue(SharedPreferenceUtil.WEB_OPEN_APP,uri.toString());
         }
     }
