+++ app/src/main/java/connect/ui/activity/wallet/ScanTransferActivity.java
@@ -22,10 +22,11 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 import connect.ui.activity.R;
 import connect.ui.activity.home.bean.MsgNoticeBean;
-import connect.ui.activity.wallet.support.ScanUrlAnalysisUtil;
 import connect.ui.base.BaseScanActivity;
 import connect.utils.ActivityUtil;
 import connect.utils.ProgressUtil;
+import connect.utils.scan.ResolveScanUtil;
+import connect.utils.scan.ResolveUrlUtil;
 import connect.view.ScanBgView;
 import connect.view.album.entity.ImageInfo;
 import connect.view.album.ui.activity.PhotoAlbumActivity;
@@ -47,7 +48,7 @@ public class ScanTransferActivity extends BaseScanActivity {
     @Bind(R.id.photos_tv)
     TextView photosTv;
     private ScanTransferActivity mActivity;
-    private ScanUrlAnalysisUtil analysisUtil;
+    private ResolveScanUtil resolveScanUtil;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -62,17 +63,17 @@ public class ScanTransferActivity extends BaseScanActivity {
     public void initView() {
         mActivity = this;
         setViewFind(capturePreview, captureCropView, captureContainer);
-        analysisUtil = new ScanUrlAnalysisUtil(mActivity);
+        resolveScanUtil = new ResolveScanUtil(mActivity);
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
     public void onEventMainThread(MsgNoticeBean notice) {
-        analysisUtil.showMsgTip(notice, "scan");
+        new ResolveUrlUtil(mActivity).showMsgTip(notice, true);
     }
 
     @Override
     public void scanCall(String value) {
-        analysisUtil.analysisUrl(value);
+        resolveScanUtil.analysisUrl(value);
     }
 
     @OnClick(R.id.left_img)
@@ -112,7 +113,7 @@ public class ScanTransferActivity extends BaseScanActivity {
             ProgressUtil.getInstance().dismissProgress();
             switch (msg.what){
                 case PARSE_BARCODE_SUC:
-                    analysisUtil.analysisUrl((String) msg.obj);
+                    resolveScanUtil.analysisUrl((String) msg.obj);
                     break;
             }
         }
